+++ public/class-tmsm-woocommerce-booking-thalasso-public.php
@@ -454,6 +454,7 @@ class Tmsm_Woocommerce_Booking_Thalasso_Public {
 		$dynamic_tags->register_tag( 'Elementor_Tag_BookResawebUrl' );
 		$dynamic_tags->register_tag( 'Elementor_Tag_ResawebUrlNoAccommodation' );
 		$dynamic_tags->register_tag( 'Elementor_Tag_BookRoomButtonLabel' );
+		$dynamic_tags->register_tag( 'Elementor_Tag_BookDiscoveryButtonLabel' );
 		$dynamic_tags->register_tag( 'Elementor_Tag_PackageRatesTable' );
 		$dynamic_tags->register_tag( 'Elementor_Tag_PackageRatesCards' );
 	}