+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-* Body class only on single
+### 1.1.10: October 18th, 2018
+* **New** - Elementor Dynamic Tag: Book Room Button Label
+* **Fix** - Body class only on single
 
 ### 1.1.9: October 17th, 2018
 * **Tweak** - Refactoring post/body classes